import fixed + comments added
