+++ .gitignore
@@ -8,3 +8,4 @@ venv
 
 # user files
 app_settings.csv
+videos
