+++ ui_components/components/new_project_page.py
@@ -1,6 +1,6 @@
 import streamlit as st
 from ui_components.common_methods import create_working_assets
-from repository.local_repo.csv_data import update_project_setting
+from repository.local_repo.csv_repo import update_project_setting
 from utils.media_processor.video import resize_video
 from moviepy.video.io.VideoFileClip import VideoFileClip
 import time
