+++ videos/controlnet_test/settings.csv
@@ -5,7 +5,7 @@ last_strength,0.7
 last_custom_pipeline,None
 audio,
 input_type,Video
-input_video,pexels-mikhail-nilov-6706912.mp4
+input_video,recorded_heman.mp4
 extraction_type,Regular intervals
 width,704
 height,512
